Changed fill tree method to use recursion instead of queue - removed queue.c and queue.h - introduced define_node, print_tree, and get_price (get_price still   needs refinement) - fixed Node price type to be float - added test01.out, which is stdout output shown in c_projects.pdf

+++ src/node.h
@@ -9,7 +9,7 @@ typedef struct NODE {
         COMPOSITE
     } node_type;
     union {
-        int price;
+        float price;
         struct {
             int num_children;
             struct NODE **children;
